requirements.txt
Flask==2.3.3
requests==2.31.0
python-dateutil==2.8.2
debugpy # Required for debugging
bleach==5.0.1
google-cloud-translate==3.1.0
